[MODIFY] logout api
